alterações

+++ src/Calculadora.java
@@ -9,7 +9,7 @@ public class Calculadora {
     public int multi (int a, int b){
         return  a*b;
     }
-    public int div (int a, int b){
+    public float div (float a, float b){
         return  a/b;
     }
 }
